Adjusted spell table column names, removed prefix from upcast_effect
Added an html to hasten viewing the adjusted data. The table won't fit in terminal so this improves workflow. Need to trim whitespace from all columns so that any future operations on the data will be much easier. Wasted a lot of time remembering that I had to do that.