Added correct boolean columns for concentration and ritual
Need to standardize the hour and minute abbreviations in the "duration" column. For some reason the data only came with the abbreviation in the casting time column. Going to try use multiple functions for a column on a single line instead of doing it sequentially on lines. The usage of .replace( ) on the duration column is the incorrect way but I know how to fix it and it's not hurting anything. Fixt next push. I swear.