Renamed to get_Fusion_targets( ), tested new subquery strategy.
find_Fusion_targets( ) renamed to get_Fusion_targets( ) to keep naming consistent with the Synchro query functions. The new subquery strategy has been added and tested so that a TEMPORARY VIEW is no longer necessary. Changed the return of get_Fusion_targets( ) so that each result is a tuple in the format (Fusion.*, Spell.*). This directly reflects the result of the table joins. Any data retained/tossed is up to the user.