LOTS (check description)
Should've done this as multiple commits but I needed to make the changes as they were still in my head.

Python version changed since adbc sqlite driver required the downgrade. New polars version just came out. Added adbc to try it with writing polars dataframes to the database.

Gestalt got turned into a backup so that I could rebuild it using a new method.

Added an empty Fusions csv as a stub for the new Fusion import method.

Added new method to import default Fusions since the previous method was really poor and hard coded.

Changed "Populate_Gestalt" to "Build_Gestalt" since the verbage makes more sense for what the file does. Pulled the data insertion out of the function and put it in a different file. Added "NOT NULL" constraint to "school" in "Spell" table since that was an artifact from pulling the XYZ out of that table.

"Import_Spell_Data" renamed to "Import_Spells" since I aliased it as that in the other files anyways.

New purpose for "Populate_Gestalt" where it'll be the file used to populate the newly created database. The file made so far is to be used with pandas. Another file with similar functionality will be made to work with polars.